Splitting up the route definitions from the router config and creation
